Add a rake task to insert registry mirror configuration
With caasp v4 it will be possible to configure a registry mirror during
installation in YaST.

This mirror will initially only be stored on the admin and needs to be inserted
into the database, to make it available to the salt-states.

For this purpose this rake task will allow to read a generated configuration
file and insert those mirrors into the database.